First, I added unit tests for MenuOption4. Second, I tweaked how MenuOption4 behaves when 2 files exist that have the same contents/hash. Third, I improved code reuse and readability for all unit tests.

+++ File_Integrity_Utility/ProgramFiles/MenuOptions/MenuOption4.cs
@@ -1,6 +1,6 @@
 ﻿namespace File_Integrity_Utility.ProgramFiles.MenuOptions
 {
-    class MenuOption4
+    public class MenuOption4
     {
         public static void RenameAllTopLevelFilesInGivenFolderAsTheirHash()
         {
@@ -21,6 +21,11 @@
             // We know filePath will not be null, but specify "string?" to suppress warnings:
             string? fileFolderPath = Path.GetDirectoryName(filePath);
             string fileNewPath = fileFolderPath + Path.DirectorySeparatorChar + fileNewNameWithExtension;
+            if (File.Exists(fileNewPath))
+            {
+                ConsoleTools.WriteLineToConsoleInColor("File not renamed, because another file already has this name: " + fileNewNameWithExtension, ConsoleColor.Red);
+                return;
+            }
             // Because we are not changing the destination folder, the current file will not be physically moved on the disk that it is on.
             // Instead, it will simply be renamed:
             File.Move(filePath, fileNewPath);
